refactor: Share httpToWS, test helpers, and base64 utilities
- Move httpToWS into channelproto.HTTPToWS, removing duplicate
  implementations from backend/tunnel/channel.go and
  desktop/channel_relay.go
- Extract StartEchoServer, StartWriteThenCloseServer, and ParseAddr
  into backend/internal/util/testutil/net.go, removing duplicates
  from tunnel and worker service test files
- Replace private uint8ArrayToBase64 in desktopBridge.ts with import
  from shared ~/lib/base64 module
- Fix reassembly map memory leak in tunnel channel recvLoop: clean up
  partial buffer before returning on size limit violation
